+++ app/src/main/kotlin/com/thomaskioko/tvmaniac/app/MainActivity.kt
@@ -24,7 +24,7 @@ import com.thomaskioko.tvmaniac.compose.theme.TvManiacTheme
 import com.thomaskioko.tvmaniac.core.notifications.api.NotificationManager.Companion.EXTRA_FROM_NOTIFICATION
 import com.thomaskioko.tvmaniac.core.notifications.api.NotificationManager.Companion.EXTRA_SHOW_ID
 import com.thomaskioko.tvmaniac.datastore.api.AppTheme
-import com.thomaskioko.tvmaniac.navigation.DeepLinkDestination
+import com.thomaskioko.tvmaniac.navigation.model.DeepLinkDestination
 
 public class MainActivity : ComponentActivity() {
     private lateinit var graph: ActivityGraph
