+++ app/src/main/kotlin/com/thomaskioko/tvmaniac/app/ui/RootScreen.kt
@@ -27,9 +27,9 @@ import com.thomaskioko.tvmaniac.debug.ui.DebugMenuScreen
 import com.thomaskioko.tvmaniac.episodedetail.ui.EpisodeDetailSheet
 import com.thomaskioko.tvmaniac.home.ui.HomeScreen
 import com.thomaskioko.tvmaniac.moreshows.ui.MoreShowsScreen
-import com.thomaskioko.tvmaniac.navigation.EpisodeSheetChild
 import com.thomaskioko.tvmaniac.navigation.RootPresenter
 import com.thomaskioko.tvmaniac.navigation.RootScreen
+import com.thomaskioko.tvmaniac.navigation.model.EpisodeSheetChild
 import com.thomaskioko.tvmaniac.search.ui.SearchScreen
 import com.thomaskioko.tvmaniac.seasondetails.ui.SeasonDetailsScreen
 import com.thomaskioko.tvmaniac.settings.ui.SettingsScreen
